requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==1.26.20
